+++ src/app/controllers/roomController.ts
@@ -52,6 +52,11 @@ export async function patchRoom (req: Request, res: Response, next: NextFunction
 export async function deleteRoom (req: Request, res: Response, next: NextFunction): Promise<void> {
 	logger.silly('Deleting room')
 
+	if (typeof req.body.confirm !== 'boolean' || req.body.confirm !== true) {
+		res.status(400).json({ error: 'You must confirm the deletion' })
+		return
+	}
+
 	try {
 		await RoomModel.findByIdAndDelete(req.params.id)
 		res.status(204).send()
